Added actual HTML content
Inserted text content from the provided text file. Also added stub img
tags where they are expected to appear later.

The footer uses the text as it appears in the screenshot. The anchor
tags for unsubscribe and privacy policy are missing their correct
hyperlinks. Note that the links for the "watch this clip here" links
are inferred from the related page on the actual Stan website.